Splitted ServiceResponse in interface and class implementation
